fixes directory structure
