+++ src/components/Carousel/CarouselLoggedIn.js
