+++ src/components/Landing/index.js
@@ -1,9 +1,22 @@
-import React from 'react';
 
-const LandingPage = () =>
-  <div>
-    <h1>Landing</h1>
-    <p>The Landing Page is open to everyone, even though the user isn't signed in.</p>
-  </div>
+import React, { Component } from "react";
+import { Button } from "reactstrap";
+import { Link } from "react-router-dom"
+import Carousel from "./Carousel";
+import Header from "./Header";
 
+class LandingPage extends Component {
+  render() {
+    return (
+      <div>
+        <Button><Link to="/SignUp"> SignUp </Link> </Button>
+        <Button><Link to="/Login"> Login </Link> </Button>
+        <Carousel />
+        <Button>
+          <Link to="/Billing">BUY NOW </Link>
+        </Button>
+      </div>
+    );
+  }
+}
 export default LandingPage;
