--- src/components/Login.js
@@ -1,81 +1,0 @@
-import React, { Component } from "react";
-import { Button, FormGroup, FormControl, ControlLabel } from "react-bootstrap";
-const { firebaseAuth } = require("../firebase");
-class Login extends Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      email: "",
-      password: ""
-    };
-  }
-
-  validateForm() {
-    return this.state.email.length > 0 && this.state.password.length > 0;
-  }
-
-  handleChange = event => {
-    this.setState({
-      [event.target.id]: event.target.value
-    });
-  };
-
-  handleSubmit = async event => {
-    event.preventDefault();
-   
-      const { email, password } = event.target.elements;
-      try {
-        const user = await firebaseAuth
-          // .auth()
-          .signInWithEmailAndPassword(email.value, password.value);
-        this.props.history.push("/");
-      } catch (error) {
-        //alert(error);
-      }
-    }
-  
-
-  render() {
-    return (
-      <div className="Login">
-        <form onSubmit={this.handleSubmit}>
-          <FormGroup controlId="email" bsSize="large">
-            <ControlLabel>Email</ControlLabel>
-            <FormControl
-              autoFocus
-              type="email"
-              value={this.state.email}
-              onChange={this.handleChange}
-            />
-          </FormGroup>
-          <FormGroup controlId="password" bsSize="large">
-            <ControlLabel>Password</ControlLabel>
-            <FormControl
-              value={this.state.password}
-              onChange={this.handleChange}
-              type="password"
-            />
-          </FormGroup>
-          <Button
-            block
-            bsSize="large"
-            disabled={!this.validateForm()}
-            type="submit"
-          >
-            Login
-          </Button>
-          <Button
-            block
-            bsSize="large"
-            // disabled={!this.validateForm()}
-            type="submit"
-          >
-            Use Instgram
-          </Button>
-        </form>
-      </div>
-    );
-  }
-}
-export default Login;
