+++ src/components/PhotoGallery/PhotoGallery.js
