--- src/components/SignUp.js
@@ -1,12 +1,0 @@
-import React, {Component} from 'react';
-
-
-
-class SignUp extends Component {
-render(){
-    return (
-        <p> Sign Up Page</p>
-    );
-}
-}
-    export default SignUp;
